move LoadingScreen to /views and add activity indicator state to RequestPermissionCard
